refactor(core): organize core modules and update imports
- Created new `core` package with essential modules
- Updated imports across project to use new `core` package
- Moved common utilities and classes to `core` for better organization
