+++ src/cicada/__init__.py
@@ -1,5 +1,5 @@
 # src/cicada/__init__.py
-from .common import *
+from .core import *
 from .geometry_pipeline import *
 from .describe import *
 from .coding import *
