+++ src/cicada/tools/web_reader.py
@@ -77,7 +77,7 @@ def fetch_content(
 # Example usage:
 if __name__ == "__main__":
 
-    from cicada.common.utils import setup_logging
+    from cicada.core.utils import setup_logging
 
     setup_logging()
